refactor #8: 패키지 및 코드 정리
