+++ src/main/java/org/prgms/locomocoserver/mogakkos/domain/Mogakko.java
@@ -18,6 +18,7 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.prgms.locomocoserver.global.common.BaseEntity;
+import org.prgms.locomocoserver.mogakkos.domain.mogakkotags.MogakkoTag;
 
 @Entity
 @Getter
